+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project (partially) adheres to [Semantic Versioning](https://semver.org
 - CHANGELOG.md file
 
 ### Changed
+- Prevent text truncation when writing output to a pipe
 - `distro` & `netifaces` Python dependencies are now "frozen" to allow upstream breakages
 
 ## [v4.11.0] - 2021-03-21